project/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
gunicorn==21.2.0
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==23.2
pdfkit==1.0.0
pillow==10.3.0
pycparser==2.22
PyYAML==6.0.1
requests==2.32.3
simplejson==3.19.2
sqlparse==0.4.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
